s3_management/thirdparty_deps/requirements.txt
boto3
pyyaml
validators
requests